Moving BaseHook usages to version_compat for slack
